% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolom_catagorie.R
\name{change_column_category}
\alias{change_column_category}
\title{Pas de categorie van een kolom in een tibble aan}
\usage{
change_column_category(data, column_name, new_category)
}
\arguments{
\item{data}{De tibble waarin de kolom zich bevindt.}

\item{column_name}{De naam van de kolom die moet worden aangepast.}

\item{new_category}{De nieuwe categorie voor de kolom.}
}
\value{
De tibble met de aangepaste kolomcategorie.
}
\description{
Deze functie past de categorie van een opgegeven kolom in een tibble aan naar de opgegeven nieuwe categorie.
}
\examples{
# Voorbeeld van het gebruik van de functie
data <- tibble(
  col1 = c("1", "2", "3"),
  col2 = c(4, 5, 6)
)
data <- change_column_category(data, column_name = "col1", new_category = "character")
data
}
