% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_numbers.R
\name{generate_random_numbers}
\alias{generate_random_numbers}
\title{Genereer een lijst met willekeurige getallen}
\usage{
generate_random_numbers(length)
}
\arguments{
\item{length}{De gewenste lengte van de lijst.}
}
\value{
Een lijst met willekeurige getallen.
}
\description{
Deze functie genereert een lijst met willekeurige getallen.
}
\examples{
# Voorbeeld van het gebruik van de functie
generate_random_numbers(5)
}
