% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_words.R
\name{generate_random_words}
\alias{generate_random_words}
\title{Genereer een lijst met willekeurige woorden}
\usage{
generate_random_words(length)
}
\arguments{
\item{length}{De gewenste lengte van de lijst.}
}
\value{
Een lijst met willekeurige woorden.
}
\description{
Deze functie genereert een lijst met willekeurige woorden.
}
\examples{
# Voorbeeld van het gebruik van de functie
generate_random_words(5)
}
