% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_libaries.R
\name{load_common_libraries}
\alias{load_common_libraries}
\title{Laad veelgebruikte libraries in}
\usage{
load_common_libraries()
}
\description{
Deze functie laadt een set veelgebruikte libraries in R, zodat je ze niet handmatig hoeft te laden in elk script.
}
\examples{
# Voorbeeld van het gebruik van de functie
load_common_libraries()
}
