\name{generate_random_numbers}
\alias{generate_random_numbers}
\title{Generate Random Numbers}
\usage{
generate_random_numbers(length)
}
\arguments{
\item{length}{The desired length of the random number list.}
}
\description{
Generates a list of random numbers.
}
\examples{
generate_random_numbers(5)
}
