\name{simple_summary}
\alias{simple_summary}
\title{simple Summary}
\usage{
simple_summary(data, grouping_column, summary_columns)
}
\arguments{
\item{data}{The input data frame or tibble.}
\item{grouping_column}{The column used for grouping the data.}
\item{summary_columns}{The columns to be summarized.}
}
\description{
Creates a simple summary of the input data by grouping it based on a specific column and summarizing the selected columns.
}
\examples{
custom_summary(data, grouping_column, summary_columns)
}
