% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_summary.R
\name{simple_summary}
\alias{simple_summary}
\title{Eenvoudige samenvatting van een tibble}
\usage{
simple_summary(data, grouping_column, summary_columns)
}
\arguments{
\item{data}{De tibble om samen te vatten.}

\item{grouping_column}{De naam van de kolom om op te groeperen.}

\item{summary_columns}{Een vector met de namen van de kolommen om samen te vatten.}
}
\value{
Een samenvatting van de tibble.
}
\description{
Deze functie biedt een eenvoudige samenvatting van een tibble met behulp van de opgegeven groepering en kolommen.
}
\examples{
# Voorbeeld van het gebruik van de functie
data <- tibble(
  group = c("A", "A", "B", "B", "B"),
  value1 = c(1, 2, 3, 4, 5),
  value2 = c(10, 20, 30, 40, 50)
)
summary <- simple_summary(data, grouping_column = "group", summary_columns = c("value1", "value2"))
summary
}
